chore: fix default options
