+++ apis/nucleus/src/components/listbox/ListBoxPopoverWrapper.jsx
@@ -1,15 +1,28 @@
 import React, { useState } from 'react';
+import extend from 'extend';
 import ListBoxPopover from './ListBoxPopover';
 
-export default function ListBoxPopoverWrapper({
-  app,
-  fieldIdentifier,
-  stateName,
-  element,
-  anchorOrigin,
-  transformOrigin,
-  options = {},
-}) {
+const DEFAULTS = {
+  show: true,
+  anchorOrigin: {
+    vertical: 'bottom',
+    horizontal: 'center',
+  },
+  transformOrigin: {
+    vertical: 'top',
+    horizontal: 'center',
+  },
+};
+
+export const getOptions = (usersOptions = {}) => {
+  const squashedOptions = {
+    ...DEFAULTS,
+  };
+  extend(true, squashedOptions, usersOptions);
+  return squashedOptions;
+};
+
+export default function ListBoxPopoverWrapper({ app, fieldIdentifier, stateName, element, options = {} }) {
   const [showState, setShowstate] = useState(!!options.show);
   const handleCloseShowState = () => {
     setShowstate(false);
@@ -20,8 +33,8 @@ export default function ListBoxPopoverWrapper({
       show={showState}
       app={app}
       alignTo={{ current: element }}
-      anchorOrigin={anchorOrigin}
-      transformOrigin={transformOrigin}
+      anchorOrigin={options.anchorOrigin}
+      transformOrigin={options.transformOrigin}
       close={handleCloseShowState}
       fieldName={fieldIdentifier}
       stateName={stateName}
